Remove prose-only test contracts
