+++ OfficeIMO.Html.Tests/Html.EnginePlatform.cs
@@ -1766,4 +1766,15 @@ public partial class Html {
         Assert.InRange(figureScore.Metrics["figures"], 0D, 0.99D);
         Assert.InRange(figureScore.Metrics["figure-signatures"], 0D, 0.99D);
     }
+
+    private static string FindRepositoryRoot() {
+        for (DirectoryInfo? directory = new DirectoryInfo(AppContext.BaseDirectory); directory != null; directory = directory.Parent) {
+            if (File.Exists(Path.Combine(directory.FullName, "Directory.Build.props"))
+                && Directory.Exists(Path.Combine(directory.FullName, "OfficeIMO.Html"))) {
+                return directory.FullName;
+            }
+        }
+
+        throw new DirectoryNotFoundException("Could not locate the OfficeIMO repository root.");
+    }
 }
